+++ packages/core/src/LogicFlow.tsx
@@ -1,7 +1,5 @@
 import { render, h } from 'preact';
-import { observer } from 'mobx-react';
-// import * as mobx from 'mobx';
-// import { IReactComponent } from 'mobx-react/dist/types/IReactComponent';
+import { observer } from './util/stateUtil';
 import GraphModel from './model/GraphModel';
 import Graph from './view/Graph';
 import Dnd from './view/behavior/DnD';
@@ -121,7 +119,9 @@ export default class LogicFlow {
     });
     // 附加功能初始化
     this.tool = new Tool(this);
-    this.history = new History(this.graphModel.eventCenter);
+    if (!this.options.isSilentMode && this.options.history !== false) {
+      this.history = new History(this.graphModel, options.isPropertiesChangeHistory);
+    }
     this.dnd = new Dnd({ lf: this });
     this.keyboard = new Keyboard({ lf: this, keyboard: options.keyboard });
     // 不可编辑模式没有开启，且没有关闭对齐线
@@ -913,7 +913,7 @@ export default class LogicFlow {
   /**
    * 触发监听事件
    */
-  emit(evt: string, arg: any) {
+  emit(evt: string, arg?: any) {
     this.graphModel.eventCenter.emit(evt, arg);
   }
 
@@ -988,6 +988,7 @@ export default class LogicFlow {
     }
     // * initNodeData区分是否为虚拟节点
     const fakerNodeModel = new Model({ ...nodeConfig, virtual: true }, this.graphModel);
+    fakerNodeModel.init();
     // fix #572: 由于正常节点不关心此属性，不会将其加入model。
     // 但是上面的constructor还是要传，历史原因，有人会在constructor中基于virtual属性进行额外处理。
     fakerNodeModel.virtual = true;
@@ -1034,22 +1035,22 @@ export default class LogicFlow {
     return text;
   }
 
-  renderRawData(graphRawData) {
+  renderRawData(graphRawData, isForce = false) {
     this.graphModel.graphDataToModel(formatData(graphRawData));
-    if (!this.options.isSilentMode && this.options.history !== false) {
-      this.history.watch(this.graphModel);
+    this.emit(EventType.GRAPH_RENDERED, graphRawData);
+    if (!this.isRendered || isForce) {
+      render((
+        <Graph
+          getView={this.getView}
+          tool={this.tool}
+          options={this.options}
+          dnd={this.dnd}
+          snaplineModel={this.snaplineModel}
+          graphModel={this.graphModel}
+        />
+      ), this.container);
+      this.isRendered = true;
     }
-    render((
-      <Graph
-        getView={this.getView}
-        tool={this.tool}
-        options={this.options}
-        dnd={this.dnd}
-        snaplineModel={this.snaplineModel}
-        graphModel={this.graphModel}
-      />
-    ), this.container);
-    this.emit(EventType.GRAPH_RENDERED, this.graphModel.modelToGraphData());
   }
   /**
    * 渲染图
@@ -1079,11 +1080,11 @@ export default class LogicFlow {
    * })
    * @param graphData 图数据
    */
-  render(graphData = {}) {
+  render(graphData = {}, isForce = false) {
     if (this.adapterIn) {
       graphData = this.adapterIn(graphData);
     }
-    this.renderRawData(graphData);
+    this.renderRawData(graphData, isForce);
   }
   /**
    * 内部保留方法